pyproject.toml
[project]
name = "checkr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "langdetect>=1.0.9",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "prometheus-client>=0.21.1",
    "pydantic-settings>=2.8.1",
    "python-gitlab>=5.6.0",
    "pyyaml>=6.0.2",
    "scrubadub>=2.0.1",
    "starlette>=0.46.1",
    "structlog>=25.2.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "commitizen>=4.4.1",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.9.10",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "pep621"
update_changelog_on_bump = true
